requirements.txt
attrs
bs4
fastapi
requests
sqlalchemy
urlpath
uvicorn